degrade/requirements.txt
numpy==1.11.1
Pillow==4.0.0
